Add Sympy example for finding resonant frequency
This is very much work-in-progress; comitting so that we have a backup
before I go on holiday.

This example goes through the process of substituting the resonance
condition into the CPDR to obtain a polynomial equation for the CPDR
in terms of omega.

We use SymPy for this purpose. It takes a bit of wrangling to ensure
that we end up with something in polynomial form after the substitution.

Unfortunately, any attempts at solving the polynomial equation
algebraically using SymPy are currently failing! Maybe it will have
better luck after we substitute in some numeric values, or perhaps we
will have to pass this polynomial over to NumPy which may fare better...