Revert `factor` -> `simplify`
Using sympy.factor instead of sympy.simplify produces a different form
for the final polynomial which, while mathematically equivalent, creates
an additional spurious point when plotting the cpdr.

We probably shouldn't be attempting to plot this point in the first
place (since it is at the point where the wave frequency and electron
gyrofrequency coincide) but we'll leave that for another day.

Also fixed an issue with an 'emptiness' check on a numpy array (`if
valid_k_roots` -> `if valid_k_roots.size > 0`)