Add use of astropy.coordinates.Angle class
Angles can be defined using either degrees or radians (and probably some
other stuff).

When we need to grab the actual value, we can use .rad or .deg notation
to do so. I like this; the conversion is automatic and it makes it very
explicit which form of the angle we are using.

Performed some grouping & reordering of parameter definitions while I
was at it.