Rework _generate_poly_in_k function
Specifically to switch from using the sympy .as_poly func (which is too
slow when num_particles is larger than two) to sym.Poly.from_list for
building our polynomial in k.

Similarly, removed some calls to sym.simplify which weren't strictly
necessary and could slow things down (to be monitored - will use of our
Poly continue to be performant outside of this func?...)

We may need to revisit the use of `summation` in the Stix parameters
(versus a call to `Sum` with a *delayed* call to `.doit()`) at some
point in the future. We will always need to expand the sums *eventually*
in order to numerically evaluate them; unless expanding them here proves
problematic in the future, I'm leaving things as-is.