Add test to verify Stix.jacobian vs Sympy methods.
And it works! This commit is not my tidiest work; I've hijacked parts of
the Cpdr class and the test_normalisation.py file in a way that won't be
acceptable in the long run, but I wanted to be able to demonstrate that
our new Stix funcs were 'working-as-expected' in the short term.