Add WIP code for #44.
Added a `solve_resonant_for_x` method to the `Cpdr` class that, given an
`omega` and a range for `X`, will locate any `X` in that range that
correspond to roots of the resonant cpdr (i.e. intersection points
between the cpdr and the resonance condition).

If we use the above with our upper and lower cutoffs for omega, we
should be able to determine when the number of 'resonant roots' changes
(with varying `X`) as per paragraph 32 of Glauert.

I've strung together a quick script to demonstrate this in action. Best
used in tandem with the resonant_roots.py script to visualise results.