Refactor script that plots Cunningham's figure 2
I've decoupled the X_range that is used for Glauert's norm factor from
the one that is used for Cunningham's.
For Glauert we are now using a uniform distribution
between [X_min, min(X_max, eps * sqrt(-P/S))]
while for Cunningham's we are using the data points
from his .dat files.