change tolerance test for signedness of k_par
From abs_tol to rel_tol. This required rearranging the resonance
condition equation so that we were no longer comparing to 0 (which
wouldn't work with rel_tol). Now comparing to 1.

Also updated test_cpdr.py accordingly to ensure one of the tests
continues to raise an (expected) ValueError when the tolerance is too
strict.