revert rollout of check_units
validators module (and tests) will remain in place, but given that the
speedup this produced was ~10% with type checking turned off I think it
is safer to leave quantity_input in place. The risks outweigh the
benefit in this instance, but we'll leave this here should we ever have
good reason to revisit things.