use Stix module for numpy_polynomials...
... to avoid the use of CpdrSymbolic, which we *might* be able to remove
if numpy's polynomials prove to be sufficiently performant and scalable.

As-is, this will make performance worse. This is because the Stix module
is horribly optimised: we perform many repeated calculations of Stix
params here without changing omega and repeatedly validate units, both
of which are unnecessary.

This is one part of the code where the sympy stuff is actually
well-optimised, assuming it always works. A 'flat', lambdified
expression that we can just sub X and k into is hard to improve on
(Stix's cascading function calls will likely always be a little slower).
The danger is that things like sympy's 'diff' might not scale well
beyond 2 particles. With numpy, we might also be able to make use of
tools like numba.