remove CpdrSymbolic and update tests accordingly
test_stix_1 has been modified and is no longer very good... we used to
compare sympy results vs numeric stix. Now we compare one set of numeric
stix to baseline data. We need more baseline data and more tests, but
I'd suggest we do this alongside a revamp of the Stix module.

test_meshing is now failing, because there is a `solve_resonant_for_x`
routine in meshing.py that needs reworking! This module is not in active
use and had earlier been suggested as a candidate for moving to an
'experimental' subfolder...