rm (obsolete) failing test in test_cpdr
This used unphysical values of (X, w, k). This particular part of the
test previously returned NaN values for k and was included to improve
coverage. The parts of the code that were being covered by this have
since been modified, and the way we calculate k has also changed
slightly such that these no longer produce NaN anyway! Since our inputs
and outputs are unphysical, we are unlikely to ever get reliable and
consistent results from this part of the test and it is best removed.