Quantity -> Angle for consistency
The Angle class is derived from Quantity, so any code that works for a
Quantity should continue to work when replaced by an Angle.

Using Angle gives us access to other convenience methods (e.g. like .deg
or .rad) and should probably be preferred. I'm not sure we've employed
this consistently across the whole codebase.