convert angles to float before using `Angle`
This change lets us use "inf" as an angle in the input.json file.
Without this change, `Angle("inf", u.deg)` will result in an error.
With this change, `Angle(float("inf"), u.deg)` returns an Angle of
infinite degrees.

"inf" is useful as a potential option for mlat_cutoff (i.e. if we
*don't* want to impose a cutoff; any other value will always be smaller
than inf).

This also clears up some potential ambiguity with the inputs for Angles:
we specify that users should input a float value and that this will be
interpreted as having units of degrees, but technically there was
nothing stopping a user from inputting any string-like angle param until
now. More on angle input parameters here:
https://docs.astropy.org/en/stable/api/astropy.coordinates.Angle.html