add lower_hybrid_freq property to PlasmaPoint
Using plasmapy method as we do for the plasma- and gyro-frequencies. We
calculate the lower hybrid frequency for all particles in our plasma
(including the electrons). I'm not sure we'll ever need this full info,
but it's consistent with our approach elsewhere. We will need the lower
hybrid frequency value for protons shortly.