mv freq_cutoff_params defn out of Cpdr.__init__
We previously used logic from G&H2005 to define the freq_cutoff_params
based on electron gyrofrequency scalings within Cpdr.__init__. This is
not well suited to the bounce-averaged cases (where we have many Cpdr
with varying electron gyrofrequencies as we traverse a magnetic field
line, but want to use a fixed set of frequency cutoffs).

We have now moved this logic into a new `from_gyrofrequency_params`
function in gauss.py. The `Cpdr` constructor now expects
freq_cutoff_params to be passed in directly (as a Gaussian).